Task-004: Se modifica a Integer los indicadores que no necesitan decimal

+++ src/main/java/acme/forms/Dashboard.java
@@ -11,10 +11,10 @@ public class Dashboard implements Serializable {
 	
 	private static final long	serialVersionUID	= 1L;
 
-	Double						totalNumberOfPublicTasks;
-	Double						totalNumberOfPrivateTasks;
-	Double						totalNumberOfFinishedTasks;
-	Double						totalNumberOfNonFinishedTasks;
+	Integer						totalNumberOfPublicTasks;
+	Integer						totalNumberOfPrivateTasks;
+	Integer						totalNumberOfFinishedTasks;
+	Integer						totalNumberOfNonFinishedTasks;
 	
 	Double						averageTaskExecutionPeriods;
 	Double						deviationTaskExecutionPeriods;
